// Frontend/src/app/models/informacion-general.ts
export interface InformacionGeneral {
  nombre: string;
  documento: string;
  repLegal?: string;
  documentoRte?: string;
  direccion: string;
  pais: string;
  nombreContacto: string;
  cargo: string;
  cargoAct: string;
  correo: string;
  telefono: string;
  esPersonaPoliticaExpuesta: boolean;
}